Use webrtc-models package
