+++ homeassistant/package_constraints.txt
@@ -64,6 +64,7 @@ uv==0.4.22
 voluptuous-openapi==0.0.5
 voluptuous-serialize==2.6.0
 voluptuous==0.15.2
+webrtc-models==0.0.0b2
 yarl==1.16.0
 zeroconf==0.135.0
