+++ pyproject.toml
@@ -74,6 +74,7 @@ dependencies    = [
     "voluptuous-serialize==2.6.0",
     "voluptuous-openapi==0.0.5",
     "yarl==1.16.0",
+    "webrtc-models==0.0.0b2",
 ]
 
 [project.urls]
