+++ requirements.txt
@@ -44,3 +44,4 @@ voluptuous==0.15.2
 voluptuous-serialize==2.6.0
 voluptuous-openapi==0.0.5
 yarl==1.16.0
+webrtc-models==0.0.0b2
